+++ cecocloud-module/cecocloud-module-marc/cecocloud-module-marc-logic-api/src/main/java/es/limit/cecocloud/marc/logic/api/dto/MarcatgeMobil.java
@@ -3,6 +3,7 @@
  */
 package es.limit.cecocloud.marc.logic.api.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.validation.constraints.NotNull;
@@ -25,7 +26,7 @@ public class MarcatgeMobil {
 	@NotNull
 	private Date data;
 	private Date dataCreacio;
-	private Double latitud;
-	private Double longitud;
+	private BigDecimal latitud;
+	private BigDecimal longitud;
 
 }
